+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed fields with layout-specific handle overrides being rendered and saved with the wrong handle.
+- Fixed bulk editing fields with layout-specific handles when the same field is used more than once or across layouts.
 
 ## 5.0.0-beta.1 - 2024-09-26